Exception handling - delete

+++ src/main/java/laurentino/workshop/services/UserService.java
@@ -2,8 +2,11 @@ package laurentino.workshop.services;
 
 import laurentino.workshop.entities.User;
 import laurentino.workshop.repositories.UserRepository;
+import laurentino.workshop.services.exception.DatabaseException;
 import laurentino.workshop.services.exception.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -33,7 +36,22 @@ public class UserService {
     }
 
     public void delete(Long id) {
-        repository.deleteById(id);
+        try {
+            if (repository.existsById(id)) {
+                repository.deleteById(id);
+            } else {
+                throw new ResourceNotFoundException(id);
+            }
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
+//        try {
+//            repository.deleteById(id);
+//        } catch (EmptyResultDataAccessException e) {
+//            throw new ResourceNotFoundException(id);
+//        } catch (DataIntegrityViolationException e) {
+//            throw new DatabaseException(e.getMessage());
+//        }
     }
 
     public User update(Long id, User obj) {
